[Refactor]#112 - separating files
