+++ Projects/Feature/Scenes/SetlistScene/Component/ToastMessageView.swift
@@ -1,0 +1,29 @@
+//
+//  ToastMessageView.swift
+//  Feature
+//
+//  Created by 고혜지 on 11/7/23.
+//  Copyright © 2023 com.creative8.seta. All rights reserved.
+//
+
+import SwiftUI
+
+struct ToastMessageView: View {
+  let message: String
+  
+  var body: some View {
+     Text(message)
+      .foregroundStyle(Color.settingTextBoxWhite)
+      .font(.subheadline)
+      .padding(.vertical, 15)
+      .frame(maxWidth: .infinity)
+      .background(
+        Color.toastBurn
+          .cornerRadius(27)
+      )
+  }
+}
+
+//#Preview {
+//    ToastMessageView()
+//}
